Pull worker out of supervisor
